third_party/WebKit/LayoutTests/fast/frames/sandboxed-iframe-storage-expected.txt

This test verifies that a sandboxed iframe does not have access to any session storage, local storage or database.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Most of the PASS messages appear in frames, presented by DumpRenderTree after the main window.

  
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS window.openDatabase('SandboxedIframeStorageDisallowed', '1.0', '', 1) threw exception "SecurityError: An attempt was made to break through the security policy of the user agent.".
PASS window.indexedDB.webkitGetDatabaseNames() threw exception "SecurityError: Failed to execute 'getDatabaseNames' on 'IDBFactory': access to the Indexed Database API is denied in this context.".
PASS window.indexedDB.open('foo') threw exception "SecurityError: Failed to execute 'open' on 'IDBFactory': access to the Indexed Database API is denied in this context.".
PASS window.indexedDB.deleteDatabase('foo') threw exception "SecurityError: Failed to execute 'deleteDatabase' on 'IDBFactory': access to the Indexed Database API is denied in this context.".
PASS window.localStorage threw exception "SecurityError: Access to 'localStorage' is denied for this document. The document is sandboxed and lacks the 'allow-same-origin' flag.".
PASS window.sessionStorage threw exception "SecurityError: Access to 'sessionStorage' is denied for this document. The document is sandboxed and lacks the 'allow-same-origin' flag.".
PASS document.cookie threw exception "SecurityError: Access to 'cookie' is denied for this document. The document is sandboxed and lacks the 'allow-same-origin' flag.".


--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS window.openDatabase('SandboxedIframeStorageAllowed', '1.0', '', 1) != null is true
PASS window.localStorage != null is true
PASS window.sessionStorage != null is true
PASS document.cookie != null is true
